Eliminate DebugBacktraceFast
Summary: This optimization is dead in our production use case, because there are several additional trace options that we use for all exceptions. Rip it out. Clean up code in Exception that can now assume that $trace is a vec. Note that some extensions still use CompactBacktrace, and we'll also make a CompactBacktrace as an intermediate value if we explicitly call debug_backtrace with no flags, so we can't clean up CompactBacktrace completely.

Reviewed By: ricklavoie

Differential Revision: D27932342

fbshipit-source-id: f9370cc8e63b9b3df119eab541b254262a1f5397