align legacy/direct decl parser w.r.t. coeffect contexts
Summary:
Small differences between the two parsers was becoming problematic. Here is the unified semantic:
- coeffect contexts are collection of types, gathered as an intersection
- 99% of the time, we keep the intersection, meaning we don't simplify empty / singleton intersections and keep them as is
- One exception is for function/method context where we turn singleton `Hfun_context` into `Tgeneric` rather than singleton intersection storing a `Tgeneric`.

The main reason for this "glitch" is that the direct decl parser (the one we want to keep) does that simplification, and putting back an intersection around this special case was a bit troublesome. I decided to keep this behavior to have neater code. And the legacy parser was easy to align on this one.

Reviewed By: losvald

Differential Revision: D27905178

fbshipit-source-id: f319272e4f85cf2d5e17d68d73344a07f014a6d3