Lift restriction on struct layout count
Summary:
Rework bespoke layout indices so that we can fit in 64 distinct struct "layout bytes", for a total of 64 * 256 struct layouts.

The new requirements are documented in bespoke/layout.h, where other layout bytes are hard-coded, and enforced in struct-dict.cpp. We use masking to compute the vtable index for a given array.

Reviewed By: arnabde03, colavitam

Differential Revision: D27952158

fbshipit-source-id: 2ca32293ff57f1ac2b4a36fc98a12bea2c729477