Consolidate .runifs in hphp/test/zend/good/ext/sqlite3
Summary: All tests in this directory have a .runif and all but 1 are the same. Merge all the ones that are the same into a config.runif and leave the one different one (sqlite3_22_loadextension.php.runif) as an override to it. This accomplishes the same thing with fewer files.

Differential Revision: D27961245

fbshipit-source-id: af5a5434f488723050e8d7c5039e0da844b19006