server_progress_file
Summary:
I want to stop using IPC with monitor to communicate progress. This diff is a step towards that.

This diff introduces a "server_progress_file" alongside the existing server_finale_file. I initialized/reset server_progress_file in the same file, Exit.ml, as the existing server_finale_file. That doesn't quite make sense but it was the nicest place to put it.

Reviewed By: CatherineGasnier

Differential Revision: D27927203

fbshipit-source-id: 4d94109e53676abfb5940ace8a83ee2fa8f93036