serverProgress.send also writes to file
Summary:
I want to stop using IPC with monitor to communicate progress. This diff is a step towards that.

This diff changes serverProgress so, when someone wishes to send progress, then in addition to sending progress to monitor, it also writes progress to server_progress_file. No one yet consumes that file so there will be no behavioral differences.

I made it log+continue if it fails to write to the file for any reason.

Reviewed By: CatherineGasnier

Differential Revision: D27930389

fbshipit-source-id: 356ad9f7a1272a12d1b7768d01be73a1e7ae603d