clientConnect reads progress from file
Summary:
I want to stop using IPC with monitor to communicate progress. This diff takes us most of the way.

The previous diff already made the server write progress to a file in addition to sending it to the monitor. This diff makes the client read from the file rather than doing IPC with the monitor.

Reviewed By: CatherineGasnier

Differential Revision: D27935367

fbshipit-source-id: 69b39da7ab975e20249e99c8ba391bab348fc28a