server_specific_files
Summary:
I want to stop using IPC with monitor to communicate progress. This diff is a step towards that.

I'm going to switch to using a file "progress.<PID>.json" which the server creates and writes to, and which the client will read from. No IPC; just communication through this lock-protected file.

The client doesn't know the server's PID. It's going to have to learn about the filename of the progress file.

The client already receives a message with a related filename from the server (the "finale file" used to communicate rich exit status). This diff changes that message to be a struct, not a string. That way in the next diff we'll be able to make the message contain two filenames in a typesafe way.

Reviewed By: CatherineGasnier

Differential Revision: D27927172

fbshipit-source-id: 7e943d7de44402b32f5f83fa9290a29b3608f6c0