serverProgress.send API no longer exposes monitor
Summary:
I want to stop using IPC with monitor to communicate progress. This diff is a step towards that.

The existing ServerProgress API is strongly tied to Monitor -- it has "monitor" in the function names, and uses MonitorRpc types. I wish to stop using monitor. This diff merely changes the API around slightly so it doesn't expose MonitorRpc types, and so the function names don't have the word "monitor". No behavior difference.

Reviewed By: CatherineGasnier

Differential Revision: D27929540

fbshipit-source-id: df17515855a48c783209ef526737c22efd0ae7ed