delete dead monitor-progress code
Summary:
I want to stop using IPC with monitor to communicate progress. This diff completes that project.

In earlier diffs, we changed the server to write to file as well as to monitor. We also changed the client to read from the file instead of the monitor. This diff at last removes the part of server that writes progress to monitor.

Reviewed By: CatherineGasnier

Differential Revision: D27935634

fbshipit-source-id: f84728337f87870cc3e4a71fa09c5b975420a885