Specialize iterators over struct dicts
Summary:
It's straightforward: we specialize these iterators, both in the interpreter (with a new IterNextIndex::Struct vtable index) and the JIT.

Probably the most interesting bit is that we fix a bug in maybeMakeLoggingArray that causes us to make the first array at each site a logging one.

As a follow-up, I'll optimize the per-iteration key and value lookups with new IR ops.

Reviewed By: colavitam

Differential Revision: D27969729

fbshipit-source-id: eac179ff10239fa64a7382b0d3bed013e80c9fc2