monitor uses monitor_fd_close_delay
Summary:
The monitor was using a delay to close FDs if run on mac, or was closing them instantly otherwise.

This diff changes it to use the hh.conf flag "monitor_fd_close_delay" to control the same thing.

I also changed the implementation around. It no longer uses the awful scheduler.ml API. (which originated as something for the IDE to schedule events and handle incoming requests over sockets). The scheduler.ml in particular was a poor fit, because its implementation of cleaning up any pending FDs would also introduce a 1s delay, which is unacceptable in the monitor.

Differential Revision: D27958245

fbshipit-source-id: 00e00e866471ad27f8bd9c6e3341ba8714a78f70