add checks without readonly being on
Summary:
This diff moves some checks from readonly analysis to be run even without the typechecker option. These are checks invovling calling functions that return readonly or accessing a readonly property. These need to be run even without readonly analysis, as otherwise you could leak readonlyness by not enabling the feature.

Note the TODO listed in the analysis: if you put a readonly keyword on a file that does not have the EnableUnstableFeatures attribute, and then call a function that returns readonly while casting it, this will cause an extra error to show up since readonly analysis is not enabled.

Reviewed By: madgen

Differential Revision: D27629334

fbshipit-source-id: 91f3aa3a858dfde977a2a8f6423d33dad2bb3440