handle dependent types
Summary: This diff just handles Tdependent as a type when looking at prop access. Dependent types have an upper bound on them, and we can just look directly at the list of upper bounds for any upper bound that's a Tclass. Any set of upper bounds will share a class hierarchy, so it suffices to stop at the first bound for the result. (As property readonlyness is invariant)

Reviewed By: madgen

Differential Revision: D27724820

fbshipit-source-id: 66a89252f3b76b512750f3f358e7bd8914ab34d6