remove anti-pattern 'module Tast = Aast'
Summary:
This pattern was used a few times, and was often misleading (cases where the
current Ast was in fact the Nast, not the Tast). I'm also concerned by the fact we have a `tast.ml` file but these constructs are not defined in it, so it makes code less readable/browsable overall.

Let's just remove it has it doesn't really help

Reviewed By: madgen

Differential Revision: D27994474

fbshipit-source-id: e45069d8a95e055266cface3c3f24167e128fb66