Fix skipping for and re-enable ext_dso_test/dso_test.php
Summary:
Add support for a 'file' verb in runif files; run the check in the parent so that it can gate tests whose config is only valid in the presence of additional files that may or may not be present. Allow / in runif file 'words' to assist with this (IMO still simpler than allowing/requiring quoting of words, since they still don't need any escaping).

Use this to re-enable hphp/test/slow/ext_dso_test/dso_test.php; it can now skip correctly when the extension is not present.

While we're here we can now convert 2 more skipifs to runifs.

Reviewed By: ricklavoie

Differential Revision: D28017843

fbshipit-source-id: bfa2928cdf220f163b0f9053c7bbca3b1901f519