add .err.json file to saved state generation
Summary:
Add an errors file `"<name>.err.json"` in JSON representation for the re-architecture's consumption. This should only be present in the 64-bit saved state, which is what we intend to use for the re-architecture. Aside from using a different representation, this new file also has a few other differences:
- It contains a flat list of filenames that have errors, with no information on phases.
- It will still be produced if there are no errors (as opposed to `.err`, which omits the file completely if there are no errors.).

The existing `.err` file contains an OCaml blob, which doesn't really suit the re-architecture. We should be removing that file once the re-architecture fully succeeds the current `hh_server`.

Differential Revision: D27798684

fbshipit-source-id: 1a1bf877d35c029cd39760d7d17c02cd5ab4422b