Eliminate static-multi-prop optimization
Summary: We weren't able to measure a win for this change in the release it landed in. It introduces extra complexity, like storing the initial static prop in two places and having a secondary multi-prop array with indices that don't match 1-1 with the primary array. That results in some amount of added complexity in V3 of D28014693 - it took me a little while to debug the bugs there. Without a measurable win, the complexity isn't worth it.

Reviewed By: subsistence

Differential Revision: D28042765

fbshipit-source-id: 51fb08967ffc78dd23916e97d9d5c5457b364166