Implement this as an implicit generic parameter
Summary:
Currently, the `this` type is represented specially, as one of the two cases of `Tdependent`, with an explicit upper bound inline.

In many ways, `this` behaves like an "implicit" generic parameter: it's instantiated during localization, and it has an upper bound (the enclosing class). Let's simplify its treatment by actually treating it as a generic parameter called `this` (no actual parameter can have this name). Future diffs will simplify the treatment of traits (using bounds to represent `require extends` and `require implements`) which is one step towards a `require is` feature.

One way in which it differs from generic parameters is that it represents an "exact" type - the actual type of the receiver - and code in `Typing_dependent` and elsewhere needs to treat it specially. In future we might want to generalize this.

A small complication is the treatment of `this` when invoking through an intersection type. We have a choice: to check
```
$obj : t & u
$obj ->foo(e)
```
for which `t::foo` has decl type `tf` and `u::foo` has decl type `uf`, we can either construct `tf[(t&u)/this] & uf[(t&u)/this]` or `tf[t/this] & uf[u/this]` where `t[u/this]` means "replace `this` by `u` in `t`". If `this` only appears in covariant positions in the type, then these are equivalent, but if it appears in both co- and contra-variant positions, then the two types are incomparable. So we construct both: hence the passing in of `this_ty_conjunct` to the `obj_get_*` functions.

Future work might avoid this complexity by treating the `this` as a generic parameter at the call site, instantiating it with a type variable. I *think* the intersection behaviour just falls out, but we need to ensure that the type variable is instantiated only with an exact type.

Reviewed By: vsiles

Differential Revision: D25898947

fbshipit-source-id: 7713768f9c5ec9f054430a350feb4be93d0830b5