Change type constant representation in Typing_defs and shallow decl
Summary:
This is the big one; the follow up to {D26626204 (https://github.com/facebook/hhvm/commit/76120ec6bba065c6d57051554a2c3209ece396aa)}. This changes the Decl representation of typeconsts from a bag of optional values to a very explicit separation by kinds. In doing this, I noticed a lot of strange patterns in our code and some fragments that I suspect are wrong. I've tagged them with a tracking task to clean up later. I've aimed to keep the functionality as close to the current one as possible, except in cases where there's a clear improvement.

The changes in Typing_extends are the most beneficial. The new setup is more verbose, but the various `Option.map2` calls are gone. It should now be clear exactly which operations happen for which constructs.

Reviewed By: losvald

Differential Revision: D28019526

fbshipit-source-id: 48cce4dac2e5c94cad4546c14e70ed4c742fbf35