Optimize locally_propagated_states
Summary:
When performing DCE, we need to walk backwards over a block, and know
the interp state for each instruction. To accomplish this, we use
locally_propagated_states, which interps the block, and stores the
state at every instruction. This is straight forward, but
inefficient. The inefficiency comes because we need to make a fully
copy of the state for each step. Moreover, since the state copies will
share Types, the Types will have non-zero ref-counts, which means
mutations will trigger a full copy of the specialized data). For
blocks with large states, the overhead can be quite high. In fact, the
scaling can be O(N^2).

Accomplish the same goal using a more efficient scheme. Interp the
block as usual, but don't make copies of the state. Instead mutate one
copy of the state (letting us mutate in place). In addition, record
the mutations to the state in an "undo log", which is a list of
actions which can be reversed. By partially "executing" the undo log,
the state can be "rewound" to any previous instruction. So for DCE, we
start at the final state, and rewind it step by step to get the data
we need.

In addition, turn some copying of value into const refs and a bit of
cleanup and dead code removal.

Reviewed By: mofarrell

Differential Revision: D28014772

fbshipit-source-id: b461500c7bab0e98af47b1480d37108730352059