Cache return type and in-out params on FuncFamily
Summary:
When a FuncFamily is very large, iterating over every FuncInfo to
calculate the return type can become very expensive. Likewise with
calculating the in-out params. Cache the return type on the FuncFamily
using LockFreeLazy, so we only need to do this expensive calculation
once. If the return types of any of the member FuncInfos changes, we
reset the LockFreeLazy and will calculate it again, as needed. We
don't actually need to cache the in-out params, as they never
change. Instead we calculate them once when we create the FuncFamily
and always use that. This speeds up analysis.

Differential Revision: D27991515

fbshipit-source-id: e45049a0b708e1fc44c6281e1dec73e980e3981d