Optimize emit_bytecode
Summary:
emit_bytecode() constructs a lambda for each possible bytecode, then
switches on the bytecode and calls the appropriate one. The problem
with this is that each lambda captures some variables from the
environment, and there's many different possible bytecodes. As a
result we spend a surprising amount of time simply constructing and
destroying these lambdas on the stack.

There's no real need for them. Simply expand the relevant macros
inline in the switch statement, instead of doing it in separate
lambdas.

Reviewed By: oulgen

Differential Revision: D28015890

fbshipit-source-id: 546226198780c4492c2107c2912f75c9eb93281c