Do return type refinement during analyze_class
Summary:
analyze_class() is responsible for class-at-a-time analysis. The idea
is to analyze all of the methods in a class together, which lets us
refine the class's property state to a fixed point. However it does
not do any "local" analysis for anything else. Everything else is left
to the Index, which means we need to call analyze_class() multiple
times to reach a fixed point for them (the main analyze iteratively
loop).

This can be quite inefficient in a number of cases. The worst is
involving return types. If you have a large class with many methods,
and the return type of each of those methods depends on the return
type of another method, we'll have to iterate an excessive amount of
times. The worst case is if there's a chain of such dependencies. In
that case, we'll have to do N rounds, where N is the size of the
largest chain. Since each round has a lot of setup, this is a lot of
wasted work.

Improve this by incorportating return type information refinement into
analyze_class() as well. When we infer a better return type for a
method on the current class, we'll take advantage of that return type
immediately. Moreover, we'll reschedule any users of that refined
return type, so we'll loop until we reach a fixed point. This reduces
the number of redundant analysis rounds considerably.

Reviewed By: mofarrell

Differential Revision: D28035764

fbshipit-source-id: 18987115901741a047ea2040660d5935c96b0286