Use more threads during HHBBC final pass
Summary:
We had previously reduced the number of threads during the HHBBC final
pass because the final pass produced data faster than we could commit
it to disk. As a result, the UnitEmitters pilled up and used too much
memory. Also using too many threads would steal CPU from sqlite.

None of this is an issue any longer. RepoFile has no trouble keeping
up with the final pass. By default, set the final pass thread count to
be the normal thread count minus 2. The minus 2 is to account for the
RepoFile creation thread, and for the cleanup thread.

Reviewed By: oulgen

Differential Revision: D28007744

fbshipit-source-id: 385314ab6dc69d98ce0e33c6f924dc93d1d3404c