Add support for registering dependencies on FuncFamily
Summary:
Right now we don't support registering a dependency directly on
FuncFamily. Instead we iterate over all the members of the FuncFamily
and register a dependency on them. This greatly bloats the size of the
dependency map and is slow (some FuncFamilies can get very large).

Add support for registering a dependency directly on FuncFamily. If a
FuncInfo changes, then any dependency on that func, or any FuncFamily
that func depends on, is rescheduled.

This required adding a back-link from a FuncInfo to any FuncFamily it
is a part of. This uses some memory, but it's not a lot.

Reviewed By: oulgen

Differential Revision: D27991510

fbshipit-source-id: 2e3a3d2cad1e847c764fd33c91e2c5c1d17221dc