Parallelize preresolve algorithm
Summary:
The reworking of the preresolve algorithm was a good start, but we can
go even further by parallelizing it. At each given "level" of classes
(distance from root), we can process the classes in parallel. The
algorithm, more or less, works as is. The only major change needed is
to defer any mutations to shared data until a round is completed.

Differential Revision: D27995037

fbshipit-source-id: c3339e42dadc9532e5037c4f7dba07eb5a11ede1