Improve public static prop logic
Summary:
We no longer have to deal with the "Don't know class but know prop
name" case for public static prop analysis (if we don't know the
class, we apply/union the type for all potential classes). This
simplifies a number of things and lets us improve the logic. In
addition, we can register a dependency on the prop itself, and not any
prop with the same name.

Differential Revision: D28024104

fbshipit-source-id: 8e9cc2fa8ccf8807917287e895903cbb2ee267e8