De-dup FuncFamily instances in HHBBC
Summary:
We create many duplicate FuncFamilies in HHBBC, which wastes an
enormous amount of memory. Only generate unique ones (re-use existing
one if available). In addition, many FuncFamilies only contain one
FuncInfo. This is a waste because we don't actually need a FuncFamily
for this case (we can just use the FuncInfo directly). Instead of
using a FuncFamily for this case, store the single resolution in a
separate table.

Reviewed By: oulgen

Differential Revision: D27991502

fbshipit-source-id: bfd7482a7474b6503292ac3d3548cbc776cfd951