Type-check context constants [3/?]: lift parser ban
Summary:
The type-checker (but not HHVM) currently raises a parsing error when
a bound (either upper/`as` or lower/`super`) is specified on a `const ctx`.
Now that type-checking lower bounds works correctly, there is no need
to be as conservative for context constants that produce type const
with lower bounds, so bounds on `const ctx` (parser already supports that).

Also, add a more comprehensive test that also test for exactness
when a type const's bounds are made equal; i.e., `super B as B`
should behave similarly to `= B` even when a class-ish is abstract.

Reviewed By: vassilmladenov

Differential Revision: D27370902

fbshipit-source-id: 2fd7899134c2f892a54029714f9290001111035b