record-type the monitor->server handoff message
Summary:
This is a no-change diff, just refactoring.

The monitor sends a pair of "handoff" messages to the server: (1) first it sends the tracker using normal Marshal, (2) then it uses the unix call "sendmsg" to send an FD.

This diff changes the type of (1) to be a record rather than just the tracker. This is to make the types more explicit, and to make it easier to add extra content to the message in future in a type-safe way.

Differential Revision: D27990567

fbshipit-source-id: ff95bb232114e57c3b212bfc9d61e4cac3c27ea6