Class constants can be lazy class, not class
Summary: We were producing class instead of class name in `emit_class_const` in HackC when `emit_class_pointers > 0`. This was preventing a number of HHBBC optimizations as it could not produce literals from classes. We now produce lazy classes when `emit_class_pointers == 2` as in all `::class` expressions. HHBBC is able to produce literals from lazy classes.

Reviewed By: paulbiss

Differential Revision: D27775404

fbshipit-source-id: c3d00c34303b1da08ea0885ebd48b350b353adf1