StructDict key pruning
Summary:
Implement StructDict key pruning, which enables layout selection to remove keys from KeyOrders so long as the resulting KeyOrder preserves a given threshold of the source's weight. This is done using a greedy algorithm to iteratively remove the key that invalidates the least number of KeyOrder instances for the source.

Accompanying layout selection changes will follow.

Differential Revision: D28084132

fbshipit-source-id: 6509341fc123f62d770e0705d9193db773a72a72