Verify symbol uniqueness when not running HHBBC
Summary:
We want to make it an invariant that symbols stored in the repo are
always unique and persistent. HHBBC already verifies this and sets the
appropriate Attrs. However, running HHBBC is optional, so its
currently possible to build a repo, not optimize it, then try to run
with it.

To deal with this, have HPHPc do the same uniqueness checks when not
running HHBBC. When running HHBBC we continue to let it do the checks
(which is the normal case). To ensure identical error messages, export
the uniqueness checking functions in HHBBC and have HPHPc use them.

Modify the test runner (and most of the hhbbc_assert files) to now be
hphpc_assert, since we expect the error to happen when running hphpc
(repo seperate), or when running hhbbc within hphpc (normal repo),
which is indistinguishable.

Reviewed By: WizKid

Differential Revision: D28141027

fbshipit-source-id: 046923cea6cf527792132067dbab9bc18b75f8e5