Add flag to error on concrete-concrete override
Summary:
One of the driving principles of the type constants improvement project is that there be exactly one canonical source for a given concrete type constant. In the past, Hack has allowed concrete type constants to be overridden in child classes provided that the type is the "same", though this equality property is judged via subtyping (i.e. `S <: T /\ T <: S`). However, since we can still express Decl Tanys, this falls down in the case of

```
class C {
  const type T = int;
}

class D extends C {
  const type T = Tany; // legal!
}
```

This restriction closes this loophole, and it opens up a path for the runtime to unconditionally error on concrete type constant overrides.

Reviewed By: zilberstein

Differential Revision: D28155228

fbshipit-source-id: 10bd863104bc7759e4cb984d2678c50db2911acc