Mangle Autoload.DB.Path with build information and parser flags
Summary:
Native Facts uses a command-line flag called `Autoload.DBPath`, which accepts the template placeholder `"%{schema}"` as part of the path. This schema gets filled with a hash which incorporates all the dependencies that should invalidate the DB.

The Facts parser is one such dependency, and a change to the Facts parser should invalidate an old DB.

The HHVM function `mangleUnitSha1()` will combine an existing hash with a hash of the current build and parser options, so any change to the HackC code or options it's called with will cause us to save and load Facts from a new file path.

This causes us to rebuild the DB every time HHVM is pushed. The full-rebuild experience in CLI-standalone mode is pretty terrible. That's why I'm stacking this diff atop D28016712 (https://github.com/facebook/hhvm/commit/258a07028a264304eb2423581fbc872b570b448f), which improves the CLI-standalone experience.

Reviewed By: ricklavoie

Differential Revision: D28043033

fbshipit-source-id: b70b2be7a9440a6fed922249e4c7a8fc98624284