Add fanout test for saved-state scenario
Summary:
Add a test framework for fanout unit tests.

The input is a multifile-like encoding of a repo change. Files in the base repo are prefixed with `base-`. Files in the changed repo are prefixed with `changed-`.

 ---

As a first scenario, the test runner correctly performs the logic to do a saved-state initialization scenario for the `changed` repo, using a saved-state generated for the `base` repo.

Fanout hashes are extracted from the server logs, then converted to human-readable symbol names using a map provided by hh_single_type_check.

 ---

Some basic fanout tests are included in this diff.

Reviewed By: CatherineGasnier

Differential Revision: D27891574

fbshipit-source-id: d787de8ae95a4a739670a4120ea2b4c02c922296