Log fanout on --config log_categories=fanout_information on saved-state init
Summary:
For our fanout unit tests I want to extract the result of a fanout calculation.

This diff adds logging for the result of a *saved-state* fanout calculation. If `hh_server` is started with `--config log_categories=fanout_information`, it will output the fanout as a JSON array of hashes.

Reviewed By: CatherineGasnier

Differential Revision: D27913425

fbshipit-source-id: 58d2fa5fa1d17f6d3b49a7674187e1a591770b52