Log fanout on --config log_categories=fanout_information for lazy incremental fanouts
Summary:
For our fanout unit tests I want to extract the result of a fanout calculation.

This diff adds logging for the result of an *incremental* fanout calculation. If `hh_server` is started with `--config log_categories=fanout_information`, it will output the fanout as a JSON array of hashes.

Reviewed By: CatherineGasnier

Differential Revision: D27908289

fbshipit-source-id: 8107d0b5ab527d68495c63c48cb6bf660a573854