do not emit `__construct` twice when it is the target method
Summary: If a class is a dependency of an extracted function or method, its constructor will be added as a dependency. However, when the class constructor is the _target_ method this leads to it being emitted twice. This diff ensures that the constructor (or any method) of class will only be emitted once, preferring the full target definition.

Reviewed By: andrewjkennedy

Differential Revision: D28182666

fbshipit-source-id: 9dcb71ef3297d22bb3ad1f5204ffe490a1152bc0