Add flag to treat require extends and implements as ancestors in the class check
Summary:
One section of the type constants improvement proposal dealt with treating `require extends` and `require implements` on equal footing with `extends` and `implements` to avoid situations where an abstract type constant is attempting to override a concrete one.

More generally, this presents a discrepancy between folded and shallow decl, the latter of which records `require extends/implements` as ancestors in the linearization. As such, this diff is a rehash of {D27526788}, and the flag can be used to identify and drive down WWW shallow decl errors.

Differential Revision: D28162311

fbshipit-source-id: 4769780c49b2b34e03e041ef78f63fadc855d2c0