Bespoke support for static props
Summary: We add LocationType::StaticProperty (and rename ::Property to ::InstanceProperty), then add code to instrument and optimize this location type to Class. To instrument, we just stick our code into initSProps. All initialization goes through this native call, so we don't need to futz with the JIT at all, and we've just put an extra branch there in the "bespokes disabled" case. To optimize, on call to setStaticBespokeArray(), we update the class's initial value.

Reviewed By: colavitam

Differential Revision: D28014693

fbshipit-source-id: 649b1c128a3f958e2a1ea0802239d605023c8b88