Add readonly to select collection interfaces
Summary:
This diff adds readonly to a bunch of hhi and systemlib files on safe functions that clearly don't modify the original collection or array.

There could also be readonly version of various other interfaces like map, filter, etc but that is much harder to fully implement (and could require something like polymorphism to nail down). For now though, we just replace the ones that are obviously readonly already.

Reviewed By: viratyosin

Differential Revision: D28145338

fbshipit-source-id: 1ec4683308f6998243850a8c1458bf30c86c1bff