Add regression tests for typeholes
Summary:
We have a bunch of bugs for which we use the tag `typehole` in our task tracking system. For each of these, the repro is a program that
* Hack accepts with `No errors`
* HHVM produces a fatal exception when running the code
It's helpful to collect such repro's together, and actually add them to our test suite, even though the "expected output" is not really `No errors`. When a type hole is accidentally fixed, the test will "fail", providing a ready-made regression test. At this point (or if we pro-actively go after these bugs) the test should be moved into the top-level `test/typecheck` directory or some suitable subdirectory.

This diff adds a few such tests; more to follow.

Reviewed By: madgen, mpu

Differential Revision: D28188461

fbshipit-source-id: 33c95c65b0c0817aa97e6a7b861d13ce9ea2e1c6